+++ flux_metrics_api/utils.py
@@ -3,6 +3,10 @@
 #
 # SPDX-License-Identifier: (MIT)
 
+import os
+import tempfile
+from contextlib import contextmanager
+
 
 def read_file(path):
     """
@@ -11,3 +15,34 @@ def read_file(path):
     with open(path, "r") as fd:
         content = fd.read()
     return content
+
+
+@contextmanager
+def workdir(dirname):
+    """
+    Provide context for a working directory, e.g.,
+
+    with workdir(name):
+       # do stuff
+    """
+    here = os.getcwd()
+    os.chdir(dirname)
+    try:
+        yield
+    finally:
+        os.chdir(here)
+
+
+def get_tmpdir(tmpdir=None, prefix="", create=True):
+    """
+    Get a temporary directory for an operation.
+    """
+    tmpdir = tmpdir or tempfile.gettempdir()
+    prefix = prefix or "flux-metrics-api-tmp"
+    prefix = "%s.%s" % (prefix, next(tempfile._get_candidate_names()))
+    tmpdir = os.path.join(tmpdir, prefix)
+
+    if not os.path.exists(tmpdir) and create is True:
+        os.mkdir(tmpdir)
+
+    return tmpdir