+++ CHANGELOG.md
@@ -14,5 +14,6 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/converged-computing/flux-metrics-api/tree/main) (0.0.x)
+ - On the fly metric (from a custom file) support, and job queue counts (0.0.11)
  - Support for certificates for uvicorn and change default port to 8443 (0.0.1)
  - Skelton release (0.0.0)